pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-governance-sdk"
version = "0.1.1"
description = "SDK for governance, telemetry, guardrails, and compliance for AI agents."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "Agent Governance Team" }]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
]

dependencies = [
  "pydantic>=2.5",
  "PyYAML>=6.0",
  "httpx>=0.25",
  "tenacity>=8.2",
  "jsonschema>=4.21",
  "opentelemetry-api>=1.22",
]

[project.optional-dependencies]
telemetry = [
  "google-cloud-logging>=3.9",
]

tracing = [
  "opentelemetry-sdk>=1.22",
  "opentelemetry-exporter-gcp-trace>=1.6",
  "opentelemetry-propagator-gcp>=1.6",
  "opentelemetry-instrumentation-httpx>=0.43b0",
  "opentelemetry-semantic-conventions>=0.43b0",
]

registry = [
  "google-cloud-bigquery>=3.17",
]

test = [
  "pytest>=7.4",
  "pytest-asyncio>=0.23",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
agent_governance = ["guardrails/*.yaml"]

[tool.pytest.ini_options]
addopts = "-q"
